A lot of changes to deal with the fact that many D-Bus services use variant typed arguments extensively. Inside those, we don't have any introspection information, and the other side might have expectations that we don't know about. Hence, we now have classes to explicitly identify the format we want to pass the object in.
Using these new classes, our implementation of the D-Bus menu protocol 
has is finally progressed into working state!


git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/dbuskit/trunk@37606 72102866-910b-0410-8b05-ffd578937521

+++ Bundles/DBusMenu/DKMenuProxy.h
@@ -40,7 +40,7 @@
 }
 - (id)initWithMenu: (NSMenu*)menu;
 - (void)menuUpdated: (NSMenu*)menu;
-- (NSUInteger)DBusIDForMenuObject: (NSMenuItem*)item;
+- (int32_t)DBusIDForMenuObject: (NSMenuItem*)item;
 - (BOOL)isExported;
 - (void)setExported: (BOOL)yesno;
 @end
